Use lisp rather than shell script for deducing hostname
This will be a bit more portable to everywhere you'd run Emacs, since
system-name is an Emacs built-in. (E.g. this should work on Windows, if
you were so inclined. Of course, the launchd scripts will _not_ work on
Windows, but as I am learning more Emacs Lisp I found it a worthwile
change anyway.)